Python: Allow source dists to install correctly
When a source distribution is installed, PEP 517 will cause pybind11 to be installed in a temporary location at build time.

https://martin-thoma.com/pyproject-toml/

Unfortunately when pybind11 is temporarily installled, it infers the wrong include path. This includes a workaround. Hopefully when that issue is fixed the workaround can be removed:

https://github.com/pybind/pybind11/issues/1067

I’ve also posted the workaround here:

https://github.com/pybind/python_example/issues/45